Refactor button components: move buttonVariants to helpers, update Dashboard layout, and enhance button size options
