+++ src/components/pages/dashboard.tsx
@@ -1,12 +1,7 @@
-import { CommentIcon } from "../team/Eyachir/icons";
 import { AllButton } from "../team/Udoy/allbutton";
-import { Button } from "../team/Udoy/button";
 
 export const Dashboard = () => (
-  <div>
-    <Button variant="primary" radius="full" size="sm">
-      <CommentIcon />
-    </Button>
+  <div className="bg-slate-100 min-h-screen">
     <AllButton />
   </div>
 );
